+++ R/report_htest_fisher.R
@@ -1,6 +1,6 @@
 # report_table -----------------
 
-.report_table_chi2 <- function(table_full, effsize) {
+.report_table_fisher <- function(table_full, effsize) {
   table_full <- cbind(table_full, attributes(effsize)$table)
   list(table = NULL, table_full = table_full)
 }
@@ -8,28 +8,19 @@
 
 # report_effectsize ---------------------
 
-.report_effectsize_chi2 <- function(x, table, dot_args, rules = "funder2019") {
-  if (chi2_type(x) %in% c("pearson", "probabilities")) {
-    args <- c(list(x), dot_args)
-    table <- do.call(effectsize::effectsize, args)
-    table_footer <- attributes(table)$table_footer
-    ci <- attributes(table)$ci
-    estimate <- names(table)[1]
-    rules <- ifelse(is.null(dot_args$rules), rules, dot_args$rules)
+.report_effectsize_fisher <- function(x, table, dot_args, rules = "funder2019") {
+  args <- c(list(x), dot_args)
+  table <- do.call(effectsize::effectsize, args)
+  ci <- attributes(table)$ci
+  estimate <- names(table)[1]
+  rules <- ifelse(is.null(dot_args$rules), rules, dot_args$rules)
 
-    args <- list(table, rules = rules, dot_args)
-    interpretation <- do.call(effectsize::interpret, args)$Interpretation
-    rules <- .text_effectsize(attr(attr(interpretation, "rules"), "rule_name"))
-  } else {
-    stop(insight::format_message(
-      "This test is not yet supported. Please open an issue at {.url https://github.com/easystats/report/issues}."
-    ), call. = FALSE)
-  }
+  args <- list(table, rules = rules, dot_args)
+  interpretation <- do.call(effectsize::interpret, args)$Interpretation
+  rules <- .text_effectsize(attr(attr(interpretation, "rules"), "rule_name"))
 
   if (estimate == "Cramers_v_adjusted") {
     main <- paste0("Adjusted Cramer's v = ", insight::format_value(table[[estimate]]))
-  } else if (estimate == "Fei") {
-    main <- paste0("Fei = ", insight::format_value(table[[estimate]]))
   } else if (estimate == "Tschuprows_t") {
     main <- paste0("Tschuprow's t = ", insight::format_value(table[[estimate]]))
   } else if (estimate == "Tschuprows_t_adjusted") {
@@ -63,7 +54,6 @@
   )
 
   table <- table[c(estimate, paste0(estimate, c("_CI_low", "_CI_high")))]
-  attributes(table)$table_footer <- table_footer
 
   list(
     table = table, statistics = statistics, interpretation = interpretation,
@@ -73,27 +63,13 @@
 
 # report_model ----------------------------
 
-.report_model_chi2 <- function(x, table) {
-  if (chi2_type(x) == "pearson") {
-    type <- " of independence between"
-    vars_full <- paste0(names(attributes(x$observed)$dimnames), collapse = " and ")
-  } else if (chi2_type(x) == "probabilities") {
-    type <- " / goodness of fit of "
-    dist <- ifelse(
-      grepl("non", attr(table, "table_footer"), fixed = TRUE), "a uniform distribution",
-      paste0("a distribution of [", paste0(
-        names(x$expected), ": n=", x$expected,
-        collapse = ", "
-      ), "]")
-    )
-
-    vars_full <- paste(x$data.name, "to", dist)
-  }
+.report_model_fisher <- function(x, table) {
+  vars_full <- paste0(names(attributes(x$observed)$dimnames), collapse = " and ")
 
   text <- paste0(
     trimws(x$method),
-    type,
-    paste0(" ", vars_full)
+    " testing the association between the variables of the ",
+    x$data.name, " dataset "
   )
 
   text
@@ -104,13 +80,13 @@ chi2_type <- function(x) {
     out <- "probabilities"
   } else if (grepl("Pearson", x$method, fixed = TRUE)) {
     out <- "pearson"
+  } else if (grepl("Fisher", x$method, fixed = TRUE)) {
+    out <- "fisher"
   }
   out
 }
 
-# report_parameters ----------------------------
-
-.report_parameters_chi2 <- function(table, stats, effsize, ...) {
+.report_parameters_fisher <- function(table, stats, effsize, ...) {
   text_full <- paste0(
     "statistically ",
     effectsize::interpret_p(table$p, rules = "default"),